@@ -8,7 +8,6 @@ using Microsoft.Bot.Builder;
 using System.Collections.Generic;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Dialogs.Choices;
-using Microsoft.Bot.Builder.Dialogs.Adaptive.Generators;
 namespace Microsoft.BotBuilderSamples
 {
     public class UserProfileDialog : ComponentDialog
@@ -21,9 +20,9 @@ namespace Microsoft.BotBuilderSamples
         {
             _userProfileAccessor = userState.CreateProperty<UserProfile>("UserProfile");
             // combine path for cross platform support
-            Dictionary<string, List<string>> lgFilesPerLocale = new Dictionary<string, List<string>>() {
-                {"", new List<string>() {Path.Combine(".", "Resources", "UserProfileDialog.lg")}},
-                {"fr", new List<string>() {Path.Combine(".", "Resources", "UserProfileDialog.fr-fr.lg")}}
+            Dictionary<string, string> lgFilesPerLocale = new Dictionary<string, string>() {
+                {"", Path.Combine(".", "Resources", "UserProfileDialog.lg")},
+                {"fr", Path.Combine(".", "Resources", "UserProfileDialog.fr-fr.lg")}
             };
             _lgGenerator = new MultiLingualTemplateEngine(lgFilesPerLocale);
 
@@ -116,7 +115,7 @@ namespace Microsoft.BotBuilderSamples
             }, stepContext);
 
             // We can send messages to the user at any point in the WaterfallStep.
-            await stepContext.Context.SendActivityAsync(ActivityFactory.CreateActivity(msg.ToString()), cancellationToken);
+            await stepContext.Context.SendActivityAsync(ActivityFactory.FromObject(msg.ToString()), cancellationToken);
 
             // WaterfallStep always finishes with the end of the Waterfall or with another dialog, here it is a Prompt Dialog.
             return await stepContext.PromptAsync(nameof(ConfirmPrompt), new PromptOptions {
@@ -137,7 +136,7 @@ namespace Microsoft.BotBuilderSamples
 
                 var msg = _lgGenerator.GenerateActivity("SummaryReadout", userProfile, stepContext);
 
-                await stepContext.Context.SendActivityAsync(ActivityFactory.CreateActivity(msg.ToString()), cancellationToken);
+                await stepContext.Context.SendActivityAsync(ActivityFactory.FromObject(msg.ToString()), cancellationToken);
             }
             else
             {
