@@ -11,7 +11,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Recognizers.Text.DataTypes.TimexExpression;
 using Microsoft.Bot.Builder.LanguageGeneration;
 using System.IO;
-using Microsoft.Bot.Builder.Dialogs.Adaptive.Generators;
 
 namespace Microsoft.BotBuilderSamples.Dialogs
 {
@@ -19,7 +18,7 @@ namespace Microsoft.BotBuilderSamples.Dialogs
     {
         protected readonly IConfiguration Configuration;
         protected readonly ILogger Logger;
-        protected TemplateEngine _lgEngine;
+        protected Templates _templates;
 
         public MainDialog(IConfiguration configuration, ILogger<MainDialog> logger)
             : base(nameof(MainDialog))
@@ -27,7 +26,7 @@ namespace Microsoft.BotBuilderSamples.Dialogs
             // combine path for cross platform support
             string[] paths = { ".", "Resources", "MainDialog.lg" };
             string fullPath = Path.Combine(paths);
-            _lgEngine = new TemplateEngine().AddFile(fullPath);
+            _templates = Templates.ParseFile(fullPath);
 
             Configuration = configuration;
             Logger = logger;
@@ -56,7 +55,10 @@ namespace Microsoft.BotBuilderSamples.Dialogs
             }
             else
             {
-                return await stepContext.PromptAsync(nameof(TextPrompt), new PromptOptions { Prompt = ActivityFactory.CreateActivity(_lgEngine.EvaluateTemplate("IntroPrompt").ToString()) }, cancellationToken);
+                return await stepContext.PromptAsync(nameof(TextPrompt), new PromptOptions
+                {
+                    Prompt = ActivityFactory.FromObject(_templates.Evaluate("IntroPrompt"))
+                }, cancellationToken);
             }
         }
 
@@ -89,7 +91,7 @@ namespace Microsoft.BotBuilderSamples.Dialogs
 
                 var timeProperty = new TimexProperty(result.TravelDate);
                 result.travelDateMsg = timeProperty.ToNaturalLanguage(DateTime.Now);
-                await stepContext.Context.SendActivityAsync(ActivityFactory.CreateActivity(_lgEngine.EvaluateTemplate("BookingConfirmation", result).ToString()), cancellationToken);
+                await stepContext.Context.SendActivityAsync(ActivityFactory.FromObject(_templates.Evaluate("BookingConfirmation", result)), cancellationToken);
             }
             else
             {
