@@ -43,7 +43,7 @@ namespace Microsoft.BotBuilderSamples
         protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken)
         {
             // Create the storage key for this conversation.
-            string key = $"{turnContext.Activity.ChannelId}/conversations/{turnContext.Activity.Conversation?.Id}";
+            var key = $"{turnContext.Activity.ChannelId}/conversations/{turnContext.Activity.Conversation?.Id}";
 
             // The execution sits in a loop because there might be a retry if the save operation fails.
             while (true)
@@ -55,7 +55,7 @@ namespace Microsoft.BotBuilderSamples
                 var (activities, newState) = await DialogHost.RunAsync(_dialog, turnContext.Activity, oldState, cancellationToken);
 
                 // Save the updated state associated with this key.
-                bool success = await _store.SaveAsync(key, newState, etag);
+                var success = await _store.SaveAsync(key, newState, etag);
 
                 // Following a successful save, send any outbound Activities, otherwise retry everything.
                 if (success)
@@ -63,7 +63,7 @@ namespace Microsoft.BotBuilderSamples
                     if (activities.Any())
                     {
                         // This is an actual send on the TurnContext we were given and so will actual do a send this time.
-                        await turnContext.SendActivitiesAsync(activities);
+                        await turnContext.SendActivitiesAsync(activities, cancellationToken);
                     }
 
                     break;
